# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FOODEX API
  version: 0.2.0
  description: Implementation baseline for FOODEX v1 mobile and web clients.
  license:
    name: Proprietary
servers:
  - url: /api/v1
security: []
...
paths:
...
  /checkout:
    post:
      tags: [Checkout]
      summary: Create an order from the authenticated cart
      operationId: checkout
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /orders:
    get:
      tags: [Orders]
      summary: List authenticated customer orders across the Platform Customer's Wholesale and Retail stores
      description: Platform Customers receive one unified order history across their own store domains. Legacy non-platform customers retain their existing channel scope.
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: store_id
          schema:
            type: integer
            minimum: 1
        - in: query
          name: channel
          schema:
            type: string
            enum: [b2b, b2c]
      responses:
        "200":
          description: Authenticated order collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCollection"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /orders/{order}:
    get:
      tags: [Orders]
      summary: Get an authorized customer order
      description: Platform Customers may open only an order owned by one of their own Wholesale or Retail domain profiles. Optional store/channel context must match the order.
      operationId: showOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - in: query
          name: store_id
          schema:
            type: integer
            minimum: 1
        - in: query
          name: channel
          schema:
            type: string
            enum: [b2b, b2c]
      responses:
        "200":
          description: Authorized order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /b2b/orders:
    get:
      tags: [B2B]
      summary: List B2B customer orders
      operationId: listB2bOrders
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: B2B order collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCollection"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /b2b/orders/{order}:
    get:
      tags: [B2B]
      summary: Get an authorized B2B order
      operationId: showB2bOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Authorized B2B order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    OrderId:
      in: path
      name: order
      required: true
      schema:
        type: integer
        minimum: 1
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      in: query
      name: per_page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Client-generated key preventing duplicate order creation for a retried checkout.
      schema:
        type: string
        minLength: 16
        maxLength: 100
  schemas:
...
    CheckoutRequest:
      type: object
      additionalProperties: false
      required: [store_id, address_id]
      properties:
        store_id:
          type: integer
          minimum: 1
        address_id:
          type: integer
          minimum: 1
        payment_method:
          type: [string, "null"]
        note:
          type: [string, "null"]
          maxLength: 1000
    OrderItem:
      type: object
      required: [id, product_id, sku, name, quantity, unit_price, line_total]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        sku:
          type: string
        name:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
          multipleOf: 0.001
        unit_price:
          type: number
          multipleOf: 0.001
        line_total:
          type: number
          multipleOf: 0.001
    Order:
      type: object
      required:
        [id, order_number, store_id, store, channel, status, currency, subtotal, grand_total, items, created_at]
      properties:
        id:
          type: integer
        order_number:
          type: string
        store_id:
          type: integer
        store:
          type: [object, "null"]
          properties:
            id:
              type: integer
            code:
              type: string
            name:
              type: string
            logo_url:
              type: [string, "null"]
        channel:
          type: string
          enum: [b2c, b2b]
        status:
          type: string
        currency:
          type: string
          enum: [EGP]
        subtotal:
          type: number
          multipleOf: 0.001
        discount_total:
          type: number
          multipleOf: 0.001
        delivery_total:
          type: number
          multipleOf: 0.001
        tax_total:
          type: number
          multipleOf: 0.001
        quote_id:
          type: [string, "null"]
          format: uuid
        quoted_at:
          type: [string, "null"]
          format: date-time
        price_tier_id:
          type: [integer, "null"]
        price_tier_code:
          type: [string, "null"]
        grand_total:
          type: number
          multipleOf: 0.001
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        invoice:
          type: [object, "null"]
          properties:
            id:
              type: integer
            invoice_number:
              type: string
            status:
              type: string
        created_at:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      required: [current_page, per_page, total]
      properties:
        current_page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    OrderCollection:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: [string, "null"]
        errors:
          type: [object, "null"]
          additionalProperties: true
  responses:
...
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found or not visible in authorized scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state or idempotency key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...